# Reduce allocations in scorecard_impl