--- a/.travis.yml
+++ b/.travis.yml
@@ -4,7 +4,7 @@ go:
     - "1.13.x"
 
 before_script:
-    - curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | sh -s -- -b $(go env GOPATH) v1.12.3
+    - curl -sfL https://install.goreleaser.com/github.com/golangci/golangci-lint.sh | sh -s -- -b $(go env GOPATH) v1.30.0
 
 script:
     - $(go env GOPATH)/golangci-lint run