In a large go program using load_management (a mysql proxy) I saw that scorecard.bucket accounts for over 5% of allocated objects in and accounts for over 1% of the cpu:

alloc_objects:
```
Total:     4773904    5822496 (flat, cum)  5.86%
  147            .          .           	s.bucket(tag).removeReference(tag) 
  148            .          .           } 
  149            .          .            
  150            .          .           func (s *scorecardImpl) bucket(tag Tag) *tagScores { 
  151            .          .           	// This hash was picked arbitrarily. The fastest, crudest, hash is probably ideal. 
  152            .    1048592           	h := fnv.New32a() 
  153      4773904    4773904           	_, _ = h.Write([]byte(tag)) 
  154            .          .           	return s.tagScoresBuckets[h.Sum32()%numBuckets]
```
cpu:
```
Total:        80ms      660ms (flat, cum)  1.23%
  147            .          .           	s.bucket(tag).removeReference(tag) 
  148            .          .           } 
  149            .          .            
  150            .          .           func (s *scorecardImpl) bucket(tag Tag) *tagScores { 
  151            .          .           	// This hash was picked arbitrarily. The fastest, crudest, hash is probably ideal. 
  152            .      110ms           	h := fnv.New32a() 
  153         40ms      480ms           	_, _ = h.Write([]byte(tag)) 
  154         40ms       70ms           	return s.tagScoresBuckets[h.Sum32()%numBuckets]
```
A lot of that cost comes from the fact that creating a byte-slice from a Tag escapes from the heap, but we need to create one in order to write to the Hash (which also escapes). I played around with it and couldn't come up with a way to avoid the allocation without using unsafe.Pointer. Ultimately, I came to this approach which basically lifts the fnv32a Write implementation and applies it directly to a Tag.

I also swapped the bytes.Buffer in the matchState for a strings.Builder to save the alloc when String-ing it.

Results look pretty decent:
```
benchmark                                old ns/op     new ns/op     delta
BenchmarkScorecard-4                     2889          2649          -8.31%
BenchmarkScorecardGenerate-4             71201         65820         -7.56%
BenchmarkBucket-4                        114           29.4          -74.21%
BenchmarkProdDataSetWithRelease-4        1466492       1353710       -7.69%
BenchmarkProdDataSetWithoutRelease-4     1071584       995530        -7.10%

benchmark                                old allocs     new allocs     delta
BenchmarkScorecard-4                     24             13             -45.83%
BenchmarkScorecardGenerate-4             276            273            -1.09%
BenchmarkBucket-4                        2              0              -100.00%
BenchmarkProdDataSetWithRelease-4        7322           5110           -30.21%
BenchmarkProdDataSetWithoutRelease-4     5739           5010           -12.70%

benchmark                                old bytes     new bytes     delta
BenchmarkScorecard-4                     556           428           -23.02%
BenchmarkScorecardGenerate-4             33047         32946         -0.31%
BenchmarkBucket-4                        36            0             -100.00%
BenchmarkProdDataSetWithRelease-4        247520        188592        -23.81%
BenchmarkProdDataSetWithoutRelease-4     207509        182176        -12.21%
```